First try to implement binary search by recursion
In this try I was able to identify the presence of element but the arrays I created for recursion in some cases showed false sizing. Hence I must improve the size detection procedure. I should also device a way to return the actual position as only the apparent position of element is known while searching.